// lib/api/generated/model/documentTargetType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 関連できる業務資源。
 */
export type DocumentTargetType =
  (typeof DocumentTargetType)[keyof typeof DocumentTargetType];

export const DocumentTargetType = {
  requirement: "requirement",
  task: "task",
  test_design: "test_design",
} as const;